7/17

+++ train.py
@@ -488,10 +488,9 @@ if __name__ == "__main__":
         from datasets import load_from_disk
         dataset = load_from_disk("temp_datasets/en-final").select(range(100))
         tokenizer = Total_model.return_tokenizer()
-        Total_model.to("cuda")
+        Total_model.to("cuda", dtype=torch.bfloat16)
         
         for data in dataset:
-            
             
             output,_,_ = Total_model(data['speech'], data['text'].lower())
             # print(output.shape)
